correction class session and add sessionUserService
