+++ core/Session.php
@@ -2,35 +2,12 @@
 
 namespace Core;
 
-use App\Repository\UserRepository;
-use Core\DefaultAbstract\DefaultAbstractEntity;
-
 /**
  * Class Session
  * @package Core
  */
 final class Session
 {
-    // Will contain the instance of our class.
-    private static $instance;
-    private static $hasLogged  = false;
-    private        $userLogged;
-
-    /**
-     * Request constructor.
-     * Prevents the creation of an instance
-     */
-    private function __construct()
-    {
-        $userId = static::getValue('id');
-
-        $this->userLogged  = empty($userId)
-            ? null
-            : (new userRepository())->findOneById($userId);
-
-        static::$hasLogged = null !== $this->userLogged;
-    }
-
     /**
      * recovers a value from the session
      * and returns the value or returns the default value.
@@ -56,35 +33,6 @@ final class Session
     }
 
     /**
-     * Singleton of session object to load once this method
-     * @return Session
-     */
-    public static function getInstance(): Session
-    {
-        if (!isset(self::$instance)) {
-            self::$instance = new self();
-        }
-
-        return self::$instance;
-    }
-
-    /**
-     * @return DefaultAbstractEntity|null
-     */
-    public static function getUserLogged(): ?DefaultAbstractEntity
-    {
-        return (static::getInstance())->retrieveUserLogged();
-    }
-
-    /**
-     * @return DefaultAbstractEntity|null
-     */
-    public function retrieveUserLogged(): ?DefaultAbstractEntity
-    {
-        return $this->userLogged;
-    }
-
-    /**
      * @param string $key
      * @param        $value
      *
@@ -95,12 +43,10 @@ final class Session
         $_SESSION[$key] = $value;
     }
 
-    /**
-     * @return bool
-     */
-    public static function hasLogged(): bool
+    public static function destroy()
     {
-        return Session::getValue('id', null);
+        session_unset();
+        session_destroy();
     }
 }
